+++ .github/workflows/integration-test.yml
@@ -84,7 +84,7 @@ jobs:
       
       - name: Run integration tests
         run: |
-          echo "check if elasticsearch is ready"
+          echo "check if opensearch is ready"
           curl -XGET https://localhost:9200 -u 'admin:admin' -k
           cd ./OpenSearch-Dashboards/plugins/security-dashboards-plugin
           yarn test:jest_server --coverage
